Build fix on ARM + Build fix when Qt doesn't have Qt3Support.
svn path=/trunk/KDE/kdebase/workspace/; revision=1098966

+++ desktopchangeosd.cpp
@@ -600,7 +600,7 @@ void DesktopChangeItem::paint( QPainter* painter, const QStyleOptionGraphicsItem
         {
         if( !m_arrowAnimation.isNull() &&
                 m_arrowAnimation.data()->state() == QAbstractAnimation::Running &&
-                !qFuzzyCompare(m_arrowValue, 1.0) )
+                !qFuzzyCompare(m_arrowValue, qreal(1.0)) )
             {
             QPixmap temp( icon.size() );
             temp.fill( Qt::transparent );
